requirement.txt
asgiref==3.8.1
Django==5.2
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
mysqlclient==2.2.7
phonenumbers==9.0.13
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
